Resolve some aliasing issues related to segment_id
In set_spatial_segment_id, the compiler cannot rule out the possibility
that segment_ids is aliased with mi_params->mi_cols. As a result, the
compiler needs to re-read mi_params->mi_cols at each iteration of the
loop. Furthermore, the compiler cannot simd-fy the loop by writing to
segment_ids in bulk. This leads to significant overhead on some
testsets.

This commit saves mi_params->mi_cols as a separate variable outside the
loop and uses memset to write to segment_ids in bulk.

Furthermore, this commit also changes the types of some segment_id from
int to uint8_t.

Performance:
| SPD_SET |  TESTSET   | AVG_PSNR | OVR_PSNR |  SSIM   | ENC_T |
|---------|------------|----------|----------|---------|-------|
|    5    |    rtc     | +0.000%  | +0.000%  | +0.000% | -0.2% |
|    5    | rtc_1080p  | +0.000%  | +0.000%  | +0.000% | -1.1% |
|    5    |  rtc_derf  | +0.000%  | +0.000%  | +0.000% | -0.1% |
|    5    | rtc_screen | +0.000%  | +0.000%  | +0.000% | -1.9% |
|---------|------------|----------|----------|---------|-------|
|    6    |    rtc     | +0.000%  | +0.000%  | +0.000% | -0.3% |
|    6    | rtc_1080p  | +0.000%  | +0.000%  | +0.000% | -1.1% |
|    6    |  rtc_derf  | +0.000%  | +0.000%  | +0.000% | -0.2% |
|    6    | rtc_screen | +0.000%  | +0.000%  | +0.000% | -1.9% |
|---------|------------|----------|----------|---------|-------|
|    7    |    rtc     | +0.000%  | +0.000%  | +0.000% | -0.7% |
|    7    | rtc_1080p  | +0.000%  | +0.000%  | +0.000% | -1.1% |
|    7    |  rtc_derf  | +0.000%  | +0.000%  | +0.000% | -0.5% |
|    7    | rtc_screen | +0.000%  | +0.000%  | +0.000% | -2.0% |
|---------|------------|----------|----------|---------|-------|
|    8    |    rtc     | +0.000%  | +0.000%  | +0.000% | -0.9% |
|    8    | rtc_1080p  | +0.000%  | +0.000%  | +0.000% | -1.2% |
|    8    |  rtc_derf  | +0.000%  | +0.000%  | +0.000% | -0.6% |
|    8    | rtc_screen | +0.000%  | +0.000%  | +0.000% | -2.1% |
|---------|------------|----------|----------|---------|-------|
|    9    |    rtc     | +0.000%  | +0.000%  | +0.000% | -1.2% |
|    9    | rtc_1080p  | +0.000%  | +0.000%  | +0.000% | -1.5% |
|    9    |  rtc_derf  | +0.000%  | +0.000%  | +0.000% | -1.0% |
|    9    | rtc_screen | +0.000%  | +0.000%  | +0.000% | -4.4% |
|---------|------------|----------|----------|---------|-------|
|   10    |    rtc     | +0.000%  | +0.000%  | +0.000% | -1.3% |
|   10    | rtc_1080p  | +0.000%  | +0.000%  | +0.000% | -1.6% |
|   10    |  rtc_derf  | +0.000%  | +0.000%  | +0.000% | -1.1% |
|   10    | rtc_screen | +0.000%  | +0.000%  | +0.000% | -4.7% |

+++ av1/common/seg_common.h
@@ -59,7 +59,7 @@ struct segmentation_probs {
 };
 
 static INLINE int segfeature_active(const struct segmentation *seg,
-                                    int segment_id,
+                                    uint8_t segment_id,
                                     SEG_LVL_FEATURES feature_id) {
   return seg->enabled && (seg->feature_mask[segment_id] & (1 << feature_id));
 }
