+++ av1/encoder/encodeframe_utils.c
@@ -325,10 +325,10 @@ void av1_update_state(const AV1_COMP *const cpi, ThreadData *td,
 
     if (!dry_run && !mi_addr->skip_txfm) {
       int cdf_num;
-      const int spatial_pred = av1_get_spatial_seg_pred(
+      const uint8_t spatial_pred = av1_get_spatial_seg_pred(
           cm, xd, &cdf_num, cpi->cyclic_refresh->skip_over4x4);
-      const int coded_id = av1_neg_interleave(mi_addr->segment_id, spatial_pred,
-                                              seg->last_active_segid + 1);
+      const uint8_t coded_id = av1_neg_interleave(
+          mi_addr->segment_id, spatial_pred, seg->last_active_segid + 1);
       int64_t spatial_cost = x->mode_costs.spatial_pred_cost[cdf_num][coded_id];
       td->rd_counts.seg_tmp_pred_cost[0] += spatial_cost;
 
@@ -338,7 +338,7 @@ void av1_update_state(const AV1_COMP *const cpi, ThreadData *td,
                                mi_col)
               : 0;
       const int use_tmp_pred = pred_segment_id == mi_addr->segment_id;
-      const int tmp_pred_ctx = av1_get_pred_context_seg_id(xd);
+      const uint8_t tmp_pred_ctx = av1_get_pred_context_seg_id(xd);
       td->rd_counts.seg_tmp_pred_cost[1] +=
           x->mode_costs.tmp_pred_cost[tmp_pred_ctx][use_tmp_pred];
       if (!use_tmp_pred) {
